Update vuepress configuration
* move theme configuration to @asterics-docs/vuepress-theme-docs
* move bigger parts of vuepress configuration to subfolder and separate files
