+++ src/vuepress/config/head.js
@@ -1,0 +1,41 @@
+module.exports = [
+  [
+    "link",
+    {
+      rel: "apple-touch-icon",
+      sizes: "180x180",
+      href: "/favicon/apple-touch-icon.png"
+    }
+  ],
+  [
+    "link",
+    {
+      rel: "icon",
+      type: "image/png",
+      sizes: "32x32",
+      href: "/favicon/favicon-32x32.png"
+    }
+  ],
+  [
+    "link",
+    {
+      rel: "icon",
+      type: "image/png",
+      sizes: "16x16",
+      href: "/favicon/favicon-16x16.png"
+    }
+  ],
+  ["link", { rel: "manifest", href: "/favicon/site.webmanifest" }],
+  [
+    "link",
+    {
+      rel: "mask-icon",
+      href: "/favicon/safari-pinned-tab.svg",
+      color: "#0ea1f2"
+    }
+  ],
+  ["meta", { name: "apple-mobile-web-app-title", content: "AsTeRICS Docs" }],
+  ["meta", { name: "application-name", content: "AsTeRICS Docs" }],
+  ["meta", { name: "msapplication-TileColor", content: "#2b5797" }],
+  ["meta", { name: "theme-color", content: "#ffffff" }]
+];
