+++ src/vuepress/config/plugins.js
@@ -1,0 +1,10 @@
+module.exports = {
+  "@vuepress/pwa": {
+    serviceWorker: true,
+    popupComponent: "AstericsSWUpdatePopup",
+    updatePopup: {
+      message: "New content is available.",
+      buttonText: "Refresh"
+    }
+  }
+};
