--- src/vuepress/plugins/bootstrap/index.js
@@ -1,4 +1,0 @@
-import Vue from "vue";
-import BootstrapVue from "bootstrap-vue";
-
-Vue.use(BootstrapVue);
