--- src/vuepress/styles/index.styl
@@ -1,33 +1,0 @@
-.navbar 
-    height 3.5rem
-
-.sidebar 
-    width 20rem
-
-@media screen and (max-width: 1024px)
-    .content 
-        max-width 80% !important
-        transition max-width 0.5s ease-out
-
-@media screen and (max-width: 768px)
-    .content 
-        max-width 99% !important
-
-.nav-link
-    padding: 0
-
-:focus 
-    outline-width: 3px
-    outline-style: solid
-    outline-color: $info
-
-.theme-default-content > h1:first-child {
-    padding-top: 2.5rem;
-    transition: padding 0.5s ease-out;
-}
-
-@media screen and (max-width: 768px) {
-    .theme-default-content > h1:first-child {
-        padding-top: 5rem;
-    }
-}
