--- src/vuepress/styles/main.scss
@@ -1,52 +1,0 @@
-// Define your variable overrides here
-$enable-shadows: true;
-// $enable-gradients: true;
-// responsive breakpoints
-// NOTE:
-// Changes of these breakpoints should be
-// aligned with the breakpoints defined
-// in palette.styl.
-//
-// Because of differing syntax, there is
-// currently no clean and easy solution,
-// except for this comment.
-$grid-breakpoints: (
-  xs: 0,
-  sm: 456px,
-  md: 789px,
-  lg: 999px,
-  xl: 1234px
-);
-
-$primary: #075069;
-$secondary: #42b983;
-$info: #0ea1d2;
-
-// $primary: #0ea1d2;
-// $secondary: #42b983;
-// $info: #075069;
-
-// Include Bootstrap and BootstrapVue SCSS files
-@import "~bootstrap/scss/bootstrap.scss";
-@import "~bootstrap-vue/src/index.scss";
-
-// Include animate.css
-@import "~animate.css/animate.min.css";
-
-@import url("https://fonts.googleapis.com/css?family=Open+Sans");
-@import url("https://fonts.googleapis.com/css?family=ZCOOL+KuaiLe");
-@import url("https://fonts.googleapis.com/css?family=Roboto");
-@import url("https://fonts.googleapis.com/css?family=Pacifico");
-@import url("https://fonts.googleapis.com/css?family=Quicksand");
-@import url("https://fonts.googleapis.com/css?family=Indie+Flower");
-@import url("https://fonts.googleapis.com/css?family=Creepster");
-@import url("https://fonts.googleapis.com/css?family=Ewert");
-@import url("https://fonts.googleapis.com/css?family=Mogra");
-@import url("https://fonts.googleapis.com/css?family=Gloria+Hallelujah");
-@import url("https://fonts.googleapis.com/css?family=Alfa+Slab+One");
-@import url("https://fonts.googleapis.com/css?family=Fredoka+One");
-@import url("https://fonts.googleapis.com/css?family=Viga");
-@import url("https://fonts.googleapis.com/css?family=Amaranth");
-@import url("https://fonts.googleapis.com/css?family=Carter+One");
-@import url("https://fonts.googleapis.com/css?family=Baloo");
-// @import url("https://fonts.googleapis.com/css?family=Roboto|ZCOOL+KuaiLe");
