--- src/vuepress/styles/palette.styl
@@ -1,40 +1,0 @@
-// colors
-$accentColor = #0ea1d2
-$textColor = #000000
-$borderColor = #eaecef
-$codeBgColor = #282c34
-$arrowBgColor = #ccc
-
-// layout
-$navbarHeight = 3.5rem
-$sidebarWidth = 20rem
-// $contentWidth = 740px
-// $contentWidth = 70%
-$contentWidth = 1400px
-// $contentWidth = 1040px
-
-// responsive breakpoints
-// $MQNarrow = 1250px
-// $MQMobile = 1020px
-// $MQMobileNarrow = 419px
-// ===
-// $MQNarrow = 999px
-// $MQMobile = 789px
-// $MQMobileNarrow = 456px
-
-// responsive breakpoints
-// NOTE:
-// Changes of these breakpoints should be
-// aligned with the breakpoints defined
-// in palette.styl.
-//
-// Because of differing syntax, there is
-// currently no clean and easy solution,
-// except for this comment.
-$MQNarrow = 1234px
-$MQMobile = 999px
-$MQMobileNarrow = 456px
-
-// code
-$lineNumbersWrapperWidth = 3.5rem
-$codeLang = js ts html md vue css sass scss less stylus go java c sh yaml py docker dockerfile makefile
